ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
google-generativeai==0.3.2
